+++ src/index/Vocabulary.cpp
@@ -230,12 +230,11 @@ auto Vocabulary<S, C, I>::lower_bound(std::string_view word,
 template <typename S, typename C, typename I>
 std::optional<ad_utility::GeometryInfo> Vocabulary<S, C, I>::getGeoInfo(
     IndexType idx) const {
-  // `PolymorphicVocabulary` or `SplitVocabulary` may have an underlying
-  // `GeoVocabulary`, which provides precomputed `GeometryInfo`
-  if constexpr (std::is_same_v<S, PolymorphicVocabulary> ||
-                ad_utility::isInstantiation<S, SplitVocabulary>) {
+  // For more information, see `isGeoInfoAvailable` below.
+  if constexpr (MaybeProvidesGeometryInfo<S>) {
     return vocabulary_.getUnderlyingVocabulary().getGeoInfo(idx.get());
   } else {
+    static_assert(NeverProvidesGeometryInfo<S>);
     return std::nullopt;
   }
 };
@@ -243,11 +242,13 @@ std::optional<ad_utility::GeometryInfo> Vocabulary<S, C, I>::getGeoInfo(
 // _____________________________________________________________________________
 template <typename S, typename C, typename I>
 bool Vocabulary<S, C, I>::isGeoInfoAvailable() const {
-  if constexpr (ad_utility::isInstantiation<S, SplitVocabulary> ||
-                std::is_same_v<S, PolymorphicVocabulary>) {
+  // For more information on the semantics of the concepts
+  // `MaybeProvidesGeometryInfo` and `NeverProvidesGeometryInfo`, please see
+  // their definitons in `PolymorphicVocabulary.h`.
+  if constexpr (MaybeProvidesGeometryInfo<S>) {
     return vocabulary_.getUnderlyingVocabulary().isGeoInfoAvailable();
-
   } else {
+    static_assert(NeverProvidesGeometryInfo<S>);
     return false;
   }
 };
